Fix UnitCellFilter
UnitCellFilter has been modified such that the generalized stress/strain
DOF are modulated by a Jacobian in order to ensures that these DOF scale
in the same way as the atomic positions and forces. Specifically, the
the strain Jacobian is V^(1/3) N^(1/6), where V is the unit cell volume
and N is the number of atoms in the unit cell. The stress Jacobian is
-V^(2/3) N^(-1/6). In addition, a "weight" argument has been added to
UnitCellFilter which allows the user to modulate the importance of the
unit cell DOF relative to the atomic DOFs. This affects both the step
size and the force-based convergence criterion used in ASE optimizers.

See http://dx.doi.org/10.1063/1.3684549 for more details.