ENH: vasp calculator xc explicitly sets GGA
The 'xc' options for PW91 and PBE should explicitly set the GGA
parameter. This does not match the previous behaviour as strictly as
leaving it implicit by only setting 'pp'. However, as pointed out by
@ehermes, the traditional behaviour leads to unexpected results when
gga is not explicitly set and pp is, e.g.:

    Vasp(xc='PBE', pp='LDA')

In the absence of any obvious use cases where it would be desirable for
this input to result in an LDA calculation, it seems better to be safe
and set the gga parameter.