ENH: add force_consistent argument to read_vasp_out()
Passing force_consistent=True returns the free energy, consistent with
forces, instead of energy with smearing extrapolated to zero. Default
is force_consistent=False for backwards compatibility.

This is consistent with the behaviour of the
ase.calculators.vasp.Vasp, which returns the free energy when
get_potential_energy() is called with force_consistent=True.

An alternative would be to read both energies from the OUTCAR file and
modify the SinglePointCalculator.get_potential_energy() method to add a
force_consistent argument.