ENH: Vasp calculator support for explicit kpts
The Vasp calculator already has support for writing explicit k-points in
cartesian or reciprocal coordinates by setting *kpts* to an (m x 3)
array (with equal weighting) or an (m x 4) array (with weights in fourth
column).

This commit adds support to the read_kpoints() method so that this
calculations can be restarted or imported from an existing calculation
folder.