Make the default EMT behaviour consistent with previous versions of ASE.
The EMT potential can now hande the cutoff the way it has always been done
in ASE, or the way it has always been done in older implementations (Asap,
Artwork, etc).

This commit changes the default to the old ASE default, and reverts changes
to the test suite necessitated by changing the default.