Make ase.calculators.emt.emt more consistent with Asap when asap_cutoff=True
There was still a strange inconsistency in how the cutoff was handled.
Also, changed the name of the parameter to the more descriptive
asap_cutoff, since now it changes two details in the cutoff to make it
more consistent with Asap.

The hard cutoff in EMT was handled in a pretty arbitrary way. It still
is, but now it is the same arbitrary way as Asap, causing the results
to be almost identical (when asap_cutoff=True).