ENG: add units members to __all__ & attribute access
The expected memebrs of the units module are added to
__all__ so they can be picked up by introspection and
do not appear as missing attributes.

The units dict from create_units is now a subclass that
implements accessing values with attributes so that it
can be a drop in replacement for the module with
new_units.eV and new_units['eV'] being equivalent.