Make it possible to select text in the GUI for copy-paste.
Some setup windows have a Python button giving a window with the
Python code that gives the same result as the GUI.  Unfortunately, it
is not possible to copy-and-paste the Python code, that seems almost
perfidious.  This is because the widget does not receive focus (see
http://stackoverflow.com/questions/21873195/readonly-tkinter-text-widget),
this patch fixes that for all text widgets in ase-gui.  I think it is
safe to enable it globally.