Read OUTCAR where # species in POSCAR/POTCAR differ
It is possible to run a VASP calculation where there are more species in
the POTCAR file than POSCAR file. ase/io/vasp/read_vasp_out will fail to
process an OUTCAR file when this is the case. This commit changes
the file to only attempt to read the minium number of species between
what is reported in the POTCAR file or the POSCAR.

This modification should only come into effect when there are more
'POTCAR:' lines in the file than species present in the 'ions per type'.
Otherwise, this program should run as before.