ENH: Skip QE configs with no results
When parsing an pw output, skip any configurations that don't have results information, this is mostly to deal with relaxations where the final configuration is printed twice and simulations where scf has not completed. Ensures that the configuration returned by indexing [-1] will have results information. Can be optinally be made to parse all configs with a keyword argument.