BUG: results-only slicing in QE reader was fixed
The previous method to slice only configurations with results in
QE output was broken as it iterated over all configurations and skipped
those with no results, this meant that index=-1 would sometimes give
a StopIteration. With this fix, it first determies which configurations
will have results and slices only from those. A test is added to ensure
the index=-1 case works properly now.