No reset of move position and proportional mouse movement
When moving a structure around the last position of the mouse is only
saved for 0.5 seconds, this means that the structure does not jump
back to the last position relative to the cursor when one has just
moved the mouse.

How much the structure is moved is now scaled relative to the window size.