API: Changed vasp2.py to vasp.py
Made changed to vasp FileIOCalculator, to ensure that it passes the existing VASP tests,
and is compatible with the ASE optimizers.

Moved VaspChargeDensity, VaspDos and xdat2traj into vasp_auxilary.py, and updated
__init__.py to reflect this change.

Made calculate() update the passed atoms, in order to ensure compatibility with the ASE optimizers.

Moved around a few function names, to avoid overwriting existing functionalities and syntax',
such as get_stress() and get_dipole_moment().

Check that the atoms.cell() is non-zero.