Fix convergence issue in niggli_reduce
A tolerance value, epsilon, is used in niggli_reduce whenever two
floating point numbers are compared. We use the suggested value of 1e-5
* V**1/3 by default. However, at one point we check to see whether
G.x * G.y * G.z is above zero; if two of those values are small (but
greater than epsilon), the product can end up between zero and epsilon.
This causes a runtime error in certain cases. To fix this, we change
the value of epsilon used in that particular check to be the more strict
epsilon**3. This also allows us to re-enable one particularly
troublesome test case.