ENH: Better standards with fall back for pdb
Modify the parsing of pdb files so that it uses fixed indexes
for all the parts of 'ATOMS' line. Provide a fallback for when
the element is not in the correct column to allow packmol and
other files to be read.

Add a test to include some exotic ATOM lines.