BUG: Fix off-by-one error in ase.io.dftb.
As per the thread [here](https://mailman.zfn.uni-bremen.de/pipermail/dftb-plus-user/2010/000493.html) the columns containing the velocities in AA/ps in the .xyz file generated by DFTB+ are the 6th, 7th and 8th. Translating that into 0-indexing arrays, these should be the 5th, 6th and 7th elements for the x, y, and z components of the velocity.