CI: Use Python images for CI tests
Run tests using the offical Python images on the CI runners.
Ensures an up to date version of pip and allows packages
to be installed from wheels.

Speed up the testing by allowing the docs testing to run
in a parallel runner. Also added a [docs] target for
installing ase that pulls in all the required Sphinx deps.