ENH: Add new xyz key_value string parser
Reaplce the regex based string parser with a character-by-character one. Has the added benefit of parsing a much more complex veriety of strings including unicode characters, nested quotes, excaped values and will not silently truncate on error.

Parser is about 50% slower, but that only applies to one line per frame, and the old parser can be passed as an argument to `io.read` to revert to the old behaviour.