BUG: xyz parser fixed for sci floats with no '.'
    'not_array="1.2 3.4 text" '

Neither the old or new xyz parsers would properly interpret floats that
do not include a decimal point e.g 6e-08. Old parser truncates to int
new parser leaves it as a string.

Change the parser to use numpy directy to convert to int or float
and add tests to ensure that this works for the buggy cases.