MAINT: Cleanup in Vasp2 code
Reduced the numebr of times we read anything from the OUTCAR file.
Some things aren't read or written from/in the vasprun.xml file,
so we still need to get some data from the OUTCAR, but try to keep it
to a minimum, as it's a lot of text parsing.

Updated the logic for reading magnetism