ENH: support for fractional/miced occupancies
The spacegroup module can now handle fractional and more importantly
mixed occupancies. That is, the second (or further) species are no
longer ignored or replaced. Instead, if occupancies < 1.0 are given,
several different species can be put on identical positions. Default
behavior to the outside is the same as before. If no occupancies are
passed, nothing changes.