MAINT: remove dodgy try-except clause
Rather than changing the 'results_required' flag after catching
a StopIteration, it is more sane to also consider a .pwo file
containing _PW_BANDS to be coming from a successful calculation
(as in the case of a finished 'nscf' run, where e.g. no energies,
forces, or fermi levels are printed out).