ENH: write (shifted) MP-grid instead of all kpts
Previously, when specifying kpts arguments of the type
{'density':4., gamma:True, ...} (i.e. not containing the
'path' keyword), the full (non-reduced) k-point mesh is
written to the input file, and QuantumEspresso does not
reduce that via some automatic symmetry detection (even
time-reversal symmetry is not applied). To make QE run
with the smaller symmetry-reduced k-points set, we just
need to provide the MP mesh and offsets instead.