BUG: fix error in parsing eigenvalues
Previously, a linebreak was assumed between the last row of
eigenvalues for one k-point and the first row of eigenvalues
of the next k-point. This is only correct if the number of
bands per k-point is a multiple of the number of columns.
In general, the switch from one k-point to the next may happen
also somewhere within a row.